+++ pkg/mid/auth_test.go
@@ -4,11 +4,13 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+
+	"github.com/go-goim/core/pkg/types"
 )
 
 func TestNewJwtToken(t *testing.T) {
 	var uid int64 = 1
-	jwt, err := NewJwtToken(uid)
+	jwt, err := NewJwtToken(types.ID(uid))
 	assert.Nil(t, err)
 	assert.NotEmpty(t, jwt)
