+++ pkg/types/default_node.go
@@ -1,12 +1,16 @@
-package snowflake
+package types
+
+import (
+	"github.com/go-goim/core/pkg/types/snowflake"
+)
 
 var (
-	defaultNode *Node
+	defaultNode *snowflake.Node
 )
 
 func SetDefaultNode(nodeBit int64) {
 	var err error
-	defaultNode, err = NewNode(nodeBit)
+	defaultNode, err = snowflake.NewNode(nodeBit)
 	if err != nil {
 		panic(err)
 	}
@@ -16,10 +20,4 @@ func assertDefaultNode() {
 	if defaultNode == nil {
 		panic("default node is not set")
 	}
-}
-
-// Generate returns a snowflake ID
-func Generate() ID {
-	assertDefaultNode()
-	return defaultNode.Generate()
 }
