+++ pkg/types/id_test.go
@@ -1,80 +1,39 @@
 package types
 
 import (
-	"reflect"
 	"testing"
+
+	"github.com/go-goim/core/pkg/types/snowflake"
 )
 
-//nolint:scopelint
-func TestID_Marshal(t *testing.T) {
-	type fields struct {
-		ID int64
+func TestID_Fromats(t *testing.T) {
+	n, err := snowflake.NewNode(1)
+	if err != nil {
+		t.Fatal(err)
 	}
-	tests := []struct {
-		name    string
-		fields  fields
-		want    []byte
-		wantErr bool
-	}{
-		{
-			name: "test1",
-			fields: fields{
-				ID: 70579728276262912,
-			},
-			want:    []byte(`"av8FMdRdcb"`),
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		id := NewID(tt.fields.ID)
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := id.MarshalJSON()
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Marshal() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("Marshal() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
 
-//nolint:scopelint
-func TestID_Unmarshal(t *testing.T) {
-	type fields struct {
-		id *ID
+	id := ID(n.Generate())
+	if id == 0 {
+		t.Fatal("id should not be 0")
 	}
-	type args struct {
-		data []byte
+
+	t.Log(id.String())
+	t.Log(snowflake.TimeFromID(id.Int64()))
+	t.Log("base2", id.Base2())
+	t.Log("base8", id.Base8())
+	t.Log("base16", id.Base16())
+	t.Log("base32", id.Base32())
+	t.Log("base36", id.Base36())
+	t.Log("base58", id.Base58())
+	t.Log("base64", id.Base64())
+	id2, err := ParseBase58([]byte(id.Base58()))
+	if err != nil {
+		t.Fatal(err)
 	}
-	tests := []struct {
-		name      string
-		fields    fields
-		args      args
-		wantInt64 int64
-		wantErr   bool
-	}{
-		{
-			name: "test1",
-			fields: fields{
-				id: &ID{},
-			},
-			args: args{
-				data: []byte(`"av8FMdRdcb"`),
-			},
-			wantInt64: 70579728276262912,
-			wantErr:   false,
-		},
+	if id != id2 {
+		t.Fatal("id should be equal to id2")
 	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := tt.fields.id.UnmarshalJSON(tt.args.data); (err != nil) != tt.wantErr {
-				t.Errorf("Unmarshal() error = %v, wantErr %v", err, tt.wantErr)
-			}
-			if tt.fields.id.Int64() != tt.wantInt64 {
-				t.Errorf("Unmarshal() got = %v, want %v", tt.fields.id.Int64(), tt.wantInt64)
-			}
-		})
+	for i := 0; i < 10; i++ {
+		t.Log(n.Generate())
 	}
 }
