+++ pkg/util/session.go
@@ -3,8 +3,10 @@ package util
 import (
 	"fmt"
 	"strconv"
+	"strings"
 
 	messagev1 "github.com/go-goim/api/message/v1"
+	"github.com/go-goim/core/pkg/types"
 )
 
 // Session generate session id
@@ -15,7 +17,7 @@ import (
 // 	010groupID00000000
 // }
 // to is groupID when type is groupChat
-func Session(tp int32, from, to int64) string {
+func Session(tp messagev1.SessionType, from, to types.ID) string {
 	// check if tp is valid
 	if tp > 0xFF || tp < 0 {
 		return ""
@@ -23,28 +25,33 @@ func Session(tp int32, from, to int64) string {
 
 	// int32 convert to uin8
 	tp &= 0xFF
-	u8 := uint8(tp)
+	u8 := strconv.FormatInt(int64(tp), 16)
 
-	switch messagev1.SessionType(tp) {
+	switch tp {
 	case messagev1.SessionType_SingleChat:
 		if from > to {
 			from, to = to, from
 		}
-		return fmt.Sprintf("%03d%020d%020d", u8, from, to)
+		return fmt.Sprintf("%02s%011s%011s", u8, from.Base58(), to.Base58())
 	case messagev1.SessionType_GroupChat:
-		return fmt.Sprintf("%03d%020d%020d", u8, to, 0)
+		return fmt.Sprintf("%02s%011s%011d", u8, from.Base58(), 0)
 	case messagev1.SessionType_Broadcast:
-		return fmt.Sprintf("%03d%020d%020d", u8, 0, 0)
+		return fmt.Sprintf("%02s%011d%011d", u8, 0, 0)
 	case messagev1.SessionType_Channel:
 		// channel session id is same as single chat
 		// one of from and to is channel id
 		if from > to {
 			from, to = to, from
 		}
-		return fmt.Sprintf("%03d%020d%020d", u8, from, to)
+		return fmt.Sprintf("%02s%011s%011s", u8, from.Base58(), to.Base58())
+	default:
+		// in default case, use single chat rule
+		// only return empty when tp is invalid
+		if from > to {
+			from, to = to, from
+		}
+		return fmt.Sprintf("%02s%011s%011s", u8, from.Base58(), to.Base58())
 	}
-
-	return ""
 }
 
 var (
@@ -54,15 +61,15 @@ var (
 	ErrInvalidSessionIDLength = fmt.Errorf("invalid session id length")
 )
 
-func ParseSession(s string) (tp int32, from, to int64, err error) {
+func ParseSession(s string) (tp int32, from, to types.ID, err error) {
 	// check if s is valid
-	if len(s) < 3+2*20 {
+	if len(s) < 2+2*11 {
 		return 0, 0, 0, ErrInvalidSessionIDLength
 	}
 
-	// first 3 bytes is session type
-	tpStr := s[:3]
-	i64, err := strconv.ParseInt(tpStr, 10, 32)
+	// first 2 bytes is session type
+	tpStr := s[:2]
+	i64, err := strconv.ParseInt(tpStr, 16, 32)
 	if err != nil {
 		return 0, 0, 0, err
 	}
@@ -74,13 +81,16 @@ func ParseSession(s string) (tp int32, from, to int64, err error) {
 	}
 
 	// get from and to
-	fromStr := s[3 : 3+20]
-	toStr := s[3+20 : 3+2*20]
-	from, err = strconv.ParseInt(fromStr, 10, 64)
+	// trim 0 from string from and to
+	// in some case, the first char is 0, so we need to trim it
+	fromStr := strings.TrimLeft(s[2:2+11], "0")
+	toStr := strings.TrimLeft(s[2+11:2+2*11], "0")
+	fmt.Println(fromStr, toStr)
+	from, err = types.ParseBase58([]byte(fromStr))
 	if err != nil {
 		return 0, 0, 0, err
 	}
-	to, err = strconv.ParseInt(toStr, 10, 64)
+	to, err = types.ParseBase58([]byte(toStr))
 	if err != nil {
 		return 0, 0, 0, err
 	}
